# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spend Analyzer API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_transactions_get",
        "parameters": [
          {
            "name": "user_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Ids"
            }
          },
          {
            "name": "account_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Ids"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "category_keys",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Keys"
            }
          },
          {
            "name": "subcategory_keys",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subcategory Keys"
            }
          },
          {
            "name": "amount_min_minor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Amount Min Minor"
            }
          },
          {
            "name": "amount_max_minor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Amount Max Minor"
            }
          },
          {
            "name": "kinds",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Kind"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kinds"
            }
          },
          {
            "name": "include_non_spend",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Non Spend"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Currency"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Patch Transaction",
        "operationId": "patch_transaction_api_transactions__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/bulk-update": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Bulk Update Transactions",
        "operationId": "bulk_update_transactions_api_transactions_bulk_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/issuers": {
      "get": {
        "tags": [
          "issuers"
        ],
        "summary": "List Issuers",
        "operationId": "list_issuers_api_issuers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Issuer"
                  },
                  "type": "array",
                  "title": "Response List Issuers Api Issuers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "issuers"
        ],
        "summary": "Create Issuer",
        "operationId": "create_issuer_api_issuers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issuer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/issuers/{issuer_id}": {
      "patch": {
        "tags": [
          "issuers"
        ],
        "summary": "Update Issuer",
        "operationId": "update_issuer_api_issuers__issuer_id__patch",
        "parameters": [
          {
            "name": "issuer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issuer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issuer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "issuers"
        ],
        "summary": "Delete Issuer",
        "operationId": "delete_issuer_api_issuers__issuer_id__delete",
        "parameters": [
          {
            "name": "issuer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issuer Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkUpdateRequest": {
        "properties": {
          "transaction_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Transaction Ids"
          },
          "category_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Key"
          },
          "subcategory_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory Key"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Kind"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "transaction_ids"
        ],
        "title": "BulkUpdateRequest"
      },
      "BulkUpdateResponse": {
        "properties": {
          "updated": {
            "type": "integer",
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "updated"
        ],
        "title": "BulkUpdateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Issuer": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "match_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Match Terms"
          },
          "default_spec_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Spec Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "match_terms"
        ],
        "title": "Issuer"
      },
      "IssuerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "match_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Match Terms"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IssuerCreate"
      },
      "IssuerUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "match_terms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Terms"
          },
          "default_spec_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Spec Id"
          }
        },
        "type": "object",
        "title": "IssuerUpdate"
      },
...
      "Kind": {
        "type": "string",
        "enum": [
          "purchase",
          "refund",
          "payment",
          "transfer",
          "fee",
          "interest",
          "adjustment"
        ],
        "title": "Kind",
        "description": "The seven `Kind` values (\u00a73.1), as a real enum rather than a `Literal` alias: FastAPI\nonly emits a named, reusable `components.schemas.Kind` in the OpenAPI document (which the\nfrontend's generated client references as `components[\"schemas\"][\"Kind\"]`) for an enum type\n\u2014 a `Literal` field is inlined per-field instead."
      },
...
      "Transaction": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "statement_id": {
            "type": "integer",
            "title": "Statement Id"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "posted_date": {
            "type": "string",
            "format": "date",
            "title": "Posted Date"
          },
          "transaction_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Date"
          },
          "description_clean": {
            "type": "string",
            "title": "Description Clean"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          },
          "category_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Key"
          },
          "subcategory_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory Key"
          },
          "merchant_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Key"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review"
          }
        },
        "type": "object",
        "required": [
          "id",
          "statement_id",
          "account_id",
          "posted_date",
          "transaction_date",
          "description_clean",
          "amount_minor",
          "currency",
          "kind",
          "category_key",
          "subcategory_key",
          "merchant_key",
          "notes",
          "needs_review"
        ],
        "title": "Transaction"
      },
      "TransactionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TransactionList"
      },
      "TransactionPatch": {
        "properties": {
          "category_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Key"
          },
          "subcategory_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory Key"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Kind"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "create_rule": {
            "type": "boolean",
            "title": "Create Rule",
            "default": false
          }
        },
        "type": "object",
        "title": "TransactionPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}